Day 19, in JavaScript
Difficult day again :(

Solved part 1 by myself, but had to run a solution from the Internet and compare outputs to find the last bug.

Was onto the right track for part 2, but needed a hint to finish it.

Part 1 takes ~6 seconds.

Part 2 takes ~10 seconds.

That’s with using multiple threads :)