Day 22, in Elm and SVG
Finally finished it :)

For part 2, I tried to make a general solution, but couldn’t figure it out, so I had to hard code one piece, once for the example and once for the full input. That required making a model of the cube out of paper.

Visualization: https://lydell.github.io/elm-aoc-template/2022-12-22/ab.html

(Looks cooler in Chrome.)